requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
roboflow>=1.1.0
PyYAML>=6.0
numpy>=1.24.0
requests>=2.31.0
Pillow>=10.0.0
